+++ src/angular/codemods/sourcemaps.ts
@@ -1,23 +1,25 @@
 /* eslint-disable @typescript-eslint/no-unsafe-member-access */
 /* eslint-disable @typescript-eslint/no-unsafe-assignment */
+
+// @ts-ignore - clack is ESM and TS complains about that. It works though
+import * as clack from '@clack/prompts';
 import * as path from 'path';
 import * as fs from 'fs';
+import { configureAngularSourcemapGenerationFlow } from '../../sourcemaps/tools/angular';
 
-export function addSourcemapEntryToAngularJSON(): void {
+export async function addSourcemapEntryToAngularJSON(): Promise<void> {
   const angularJsonPath = path.join(process.cwd(), 'angular.json');
-
   const angularJSONFile = fs.readFileSync(angularJsonPath, 'utf-8');
-
   const angularJson = JSON.parse(angularJSONFile);
 
   if (!angularJson) {
-    throw new Error('Could not find in angular.json in your project');
+    await configureAngularSourcemapGenerationFlow();
   }
 
   const projects = Object.keys(angularJson.projects as Record<string, unknown>);
 
   if (!projects.length) {
-    throw new Error('Could not find any projects in angular.json');
+    await configureAngularSourcemapGenerationFlow();
   }
 
   // Emit sourcemaps from all projects in angular.json
@@ -42,4 +44,8 @@ export function addSourcemapEntryToAngularJSON(): void {
   }
 
   fs.writeFileSync(angularJsonPath, JSON.stringify(angularJson, null, 2));
+
+  clack.log.info(
+    'Added sourcemap configuration to angular.json for all projects',
+  );
 }